Added stub tester for virtual_mirror_joewl

+++ catkin_ws/src/spring2016/joe-wl/image_average_joewl/src/image_average_joewl_node.py
@@ -30,7 +30,7 @@ class ImageAverage(object):
     def avgImage(self,image_msg):
 	    # Convert image message to CV float32 image:
 	    #image_cv = self.bridge.imgmsg_to_cv2(image_msg, "bgr8")
-        image_cv = cv2.imdecode(np.fromstring(image_msg.data, np.float32), cv2.CV_LOAD_IMAGE_COLOR)
+        image_cv = cv2.imdecode(np.fromstring(image_msg.data, np.uint8), cv2.CV_LOAD_IMAGE_COLOR)
         image_cv = np.float32(image_cv)
 
         # Incorporate image in running average (initialize if first image):
